fix: ensure consistent use of uint32 for mount namespace IDs
- Updated type conversions between uint32 and uint64 for mount namespace IDs across the codebase
- Fixed test cases in pkg/utils/syscalls/syscalls_test.go to work correctly on ARM64
- Ensured consistent type usage for both mount and network namespace IDs

+++ pkg/gadgets/advise/seccomp/tracer/tracer.go
@@ -167,7 +167,7 @@ func (t *Tracer) AttachContainer(container *containercollection.Container) error
 }
 
 func (t *Tracer) DetachContainer(container *containercollection.Container) error {
-	res, err := t.Peek(container.Mntns)
+	res, err := t.Peek(uint64(container.Mntns))
 	if err != nil {
 		t.containers[container] = []string{err.Error()}
 		return nil
